+++ src/FractalDensityMask.hpp
@@ -87,7 +87,7 @@ private:
   void make_fractal_grid(
       RandomGenerator &random_generator,
       CoordinateVector<> current_position = CoordinateVector<>(0.),
-      unsigned int current_level = 1) {
+      uint_fast8_t current_level = 1) {
 
     CoordinateVector<> x_level = current_position;
     x_level[0] += 2. * (random_generator.get_uniform_random_double() - 0.5) /
@@ -98,7 +98,7 @@ private:
                   std::pow(_L, current_level);
 
     if (current_level < _num_level) {
-      for (unsigned int i = 0; i < _N; ++i) {
+      for (uint_fast32_t i = 0; i < _N; ++i) {
         make_fractal_grid(random_generator, x_level, current_level + 1);
       }
     } else {
@@ -138,9 +138,9 @@ private:
 
       // map the coordinates to grid indices and add a point to the
       // corresponding cell
-      const unsigned int ix = x_level.x() * _distribution.size();
-      const unsigned int iy = x_level.y() * _distribution[ix].size();
-      const unsigned int iz = x_level.z() * _distribution[ix][iy].size();
+      const uint_fast32_t ix = x_level.x() * _distribution.size();
+      const uint_fast32_t iy = x_level.y() * _distribution[ix].size();
+      const uint_fast32_t iz = x_level.z() * _distribution[ix][iy].size();
 
       // use an atomic operation to add the point, to make this method thread
       // safe
@@ -179,7 +179,7 @@ private:
      * @param index Index of the next first level block that will be constructed
      * by this job.
      */
-    inline void set_index(unsigned int index) {
+    inline void set_index(uint_fast32_t index) {
       _random_generator.set_seed(_mask._first_level_seeds[index]);
     }
 
@@ -213,10 +213,10 @@ private:
     std::vector< FractalDensityMaskConstructionJob * > _jobs;
 
     /*! @brief Current index on the first level. */
-    unsigned int _current_index;
+    uint_fast32_t _current_index;
 
     /*! @brief Total number of particles on the first level. */
-    const unsigned int _num_part;
+    const uint_fast32_t _num_part;
 
     /*! @brief Lock used to ensure safe access to the internal index counter. */
     Lock _lock;
@@ -229,11 +229,11 @@ private:
      * @param worksize Number of threads to use.
      */
     inline FractalDensityMaskConstructionJobMarket(FractalDensityMask &mask,
-                                                   int worksize)
+                                                   int_fast32_t worksize)
         : _current_index(0), _num_part(mask._N) {
 
       _jobs.reserve(worksize);
-      for (int i = 0; i < worksize; ++i) {
+      for (int_fast32_t i = 0; i < worksize; ++i) {
         _jobs.push_back(new FractalDensityMaskConstructionJob(mask));
       }
     }
@@ -244,7 +244,7 @@ private:
      *
      * @param worksize Number of parallel threads that will be used.
      */
-    inline void set_worksize(int worksize) {}
+    inline void set_worksize(int_fast32_t worksize) {}
 
     /**
      * @brief Destructor.
@@ -252,7 +252,7 @@ private:
      * Clean up memory used by jobs.
      */
     inline ~FractalDensityMaskConstructionJobMarket() {
-      for (unsigned int i = 0; i < _jobs.size(); ++i) {
+      for (size_t i = 0; i < _jobs.size(); ++i) {
         delete _jobs[i];
       }
     }
@@ -265,13 +265,13 @@ private:
      * @return Pointer to a FractalDensityMaskConstructionJob, or a null pointer
      * if no more jobs are available.
      */
-    inline FractalDensityMaskConstructionJob *get_job(int thread_id) {
+    inline FractalDensityMaskConstructionJob *get_job(int_fast32_t thread_id) {
 
       if (_current_index == _num_part) {
         return nullptr;
       }
       bool has_job = false;
-      unsigned int current_index = 0;
+      uint_fast32_t current_index = 0;
       // we can only change _current_index if the lock is locked
       // similarly, we can only take decision based on the value of
       // _current_index while the lock is locked (because otherwise another
@@ -433,12 +433,12 @@ public:
         const double ncell = it.get_ionization_variables().get_number_density();
         const double Ncell = ncell * it.get_volume();
         Ntot += Ncell;
-        const unsigned int ix = (midpoint.x() - _box.get_anchor().x()) /
-                                _box.get_sides().x() * _resolution.x();
-        const unsigned int iy = (midpoint.y() - _box.get_anchor().y()) /
-                                _box.get_sides().y() * _resolution.y();
-        const unsigned int iz = (midpoint.z() - _box.get_anchor().z()) /
-                                _box.get_sides().z() * _resolution.z();
+        const uint_fast32_t ix = (midpoint.x() - _box.get_anchor().x()) /
+                                 _box.get_sides().x() * _resolution.x();
+        const uint_fast32_t iy = (midpoint.y() - _box.get_anchor().y()) /
+                                 _box.get_sides().y() * _resolution.y();
+        const uint_fast32_t iz = (midpoint.z() - _box.get_anchor().z()) /
+                                 _box.get_sides().z() * _resolution.z();
         Nsmooth += smooth_fraction * Ncell;
         Nfractal += _fractal_fraction * Ncell * _distribution[ix][iy][iz];
       }
@@ -450,12 +450,12 @@ public:
     for (auto it = grid.begin(); it != grid.end(); ++it) {
       const CoordinateVector<> midpoint = it.get_cell_midpoint();
       if (_box.inside(midpoint)) {
-        const unsigned int ix = (midpoint.x() - _box.get_anchor().x()) /
-                                _box.get_sides().x() * _resolution.x();
-        const unsigned int iy = (midpoint.y() - _box.get_anchor().y()) /
-                                _box.get_sides().y() * _resolution.y();
-        const unsigned int iz = (midpoint.z() - _box.get_anchor().z()) /
-                                _box.get_sides().z() * _resolution.z();
+        const uint_fast32_t ix = (midpoint.x() - _box.get_anchor().x()) /
+                                 _box.get_sides().x() * _resolution.x();
+        const uint_fast32_t iy = (midpoint.y() - _box.get_anchor().y()) /
+                                 _box.get_sides().y() * _resolution.y();
+        const uint_fast32_t iz = (midpoint.z() - _box.get_anchor().z()) /
+                                 _box.get_sides().z() * _resolution.z();
         const double ncell = it.get_ionization_variables().get_number_density();
         const double nsmooth = smooth_fraction * ncell;
         const double nfractal = _fractal_fraction * fractal_norm * ncell *
